# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for a private event attendance and gift registry.
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Event {
  id: string;
  title: string;
  subtitle: string;
  dateLabel: string;
  dateIso: string;
  location: string;
  address: string;
  hostName: string;
  hostEmail: string;
  rsvpDeadline: string;
}

export interface Gift {
  id: string;
  name: string;
  category: string;
  categoryLabel: string;
  quantity: number;
  reservedQuantity: number;
  availableQuantity: number;
  available: boolean;
  reservedByMe: boolean;
}

export interface Participation {
  id: string;
  userId: string;
  guestName: string;
  guestEmail: string;
  attending: boolean;
  plusOne: boolean;
  /** @nullable */
  note: string | null;
  giftIds: string[];
  giftNames: string[];
  updatedAt: string;
}

export interface ParticipationInput {
  attending: boolean;
  plusOne: boolean;
  /** @nullable */
  note: string | null;
  /** @maxItems 2 */
  giftIds: string[];
}

export type GuestParticipationInput = ParticipationInput & {
  /**
     * @minLength 2
     * @maxLength 120
     */
  guestName: string;
};

export interface GuestParticipationAccess {
  accessToken: string;
  participation: Participation;
}

export interface AdminSummary {
  totalGuests: number;
  attendingCount: number;
  declinedCount: number;
  pendingCount: number;
  plusOneCount: number;
  totalGifts: number;
  reservedGifts: number;
  availableGifts: number;
  /** @nullable */
  lastUpdatedAt: string | null;
}

export interface AdminParticipation {
  id: string;
  guestName: string;
  guestEmail: string;
  attending: boolean;
  plusOne: boolean;
  /** @nullable */
  note: string | null;
  giftNames: string[];
  updatedAt: string;
}

export interface ErrorResponse {
  error: string;
}