# lib/api-zod/src/generated/types/guestParticipationAccess.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for a private event attendance and gift registry.
 * OpenAPI spec version: 0.1.0
 */
import type { Participation } from './participation';

export interface GuestParticipationAccess {
  accessToken: string;
  participation: Participation;
}