# lib/api-zod/src/generated/types/guestParticipationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for a private event attendance and gift registry.
 * OpenAPI spec version: 0.1.0
 */
import type { ParticipationInput } from './participationInput';

export type GuestParticipationInput = ParticipationInput & {
  /**
     * @minLength 2
     * @maxLength 120
     */
  guestName: string;
};